requirements.txt
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
PyMySQL==1.1.1
requests==2.32.4
